+++ src/controls/inputs.js
@@ -1,6 +1,6 @@
+import 'mapbox-gl-geocoder';
 import template from 'lodash.template';
-import debounce from 'lodash.debounce';
-import typeahead from 'suggestions';
+import isEqual from 'lodash.isequal';
 import extent from 'turf-extent';
 
 let fs = require('fs'); // substack/brfs#39
@@ -37,101 +37,63 @@ export default class Inputs {
   animateToCoordinates(mode, coords) {
     const { origin, destination } = this.store.getState();
 
-    if (origin.type && destination.type) {
+    if (origin.geometry &&
+        destination.geometry &&
+        !isEqual(origin.geometry, destination.geometry)) {
+
       // Animate map to fit bounds.
       const bb = extent({
         type: 'FeatureCollection',
         features: [origin, destination]
       });
 
-      this.map.fitBounds([[bb[0], bb[1]], [bb[2], bb[3]]], { padding: 40 });
+      this.map.fitBounds([[bb[0], bb[1]], [bb[2], bb[3]]], { padding: 80 });
     } else {
       this.map.flyTo({ center: coords });
     }
   }
 
-  nonKeyList(code) {
-    // TAB, ESC, LEFT, RIGHT, ENTER, UP, DOWN
-    return [9, 27, 37, 39, 13, 38, 40].indexOf(code) !== -1;
-  }
-
   onAdd() {
     const {
-      queryOrigin,
-      queryDestination,
-      originCoordinates,
-      destinationCoordinates,
       clearOrigin,
       clearDestination,
+      createOrigin,
+      createDestination,
       setProfile,
       reverse
     } = this.actions;
 
-    this.originInput = this.container.querySelector('.js-origin');
-    this.destinationInput = this.container.querySelector('.js-destination');
-
-    this.originClear = this.container.querySelector('.js-origin-clear');
-    this.destinationClear = this.container.querySelector('.js-destination-clear');
-
-    this.originLoading = this.container.querySelector('.js-origin-loading');
-    this.destinationLoading = this.container.querySelector('.js-destination-loading');
-
-    // Events
-    // ============================
-
-    // Origin / Destination autosuggest
-    this.originInput.addEventListener('keydown', debounce((e) => {
-      if (this.nonKeyList(e.keyCode)) return;
-      const v = e.target.value.trim();
-      if (v.length > 2) {
-        queryOrigin(v);
-      } else {
-        this.originClear.classList.remove('active');
-      }
-    }), 300);
-
-    this.originInput.addEventListener('change', () => {
-      if (this.originTypeahead.selected) {
-        const coords = this.originTypeahead.selected.center;
-        originCoordinates(coords);
-        this.animateToCoordinates('origin', coords);
-      }
+    this.originInput = new mapboxgl.Geocoder({
+      flyTo: false,
+      placeholder: 'Choose a starting place',
+      container: this.container.querySelector('#mapbox-directions-origin-input')
     });
 
-    this.originClear.addEventListener('click', () => {
-      this.originClear.classList.remove('active');
-      clearOrigin();
+    this.map.addControl(this.originInput);
+
+    this.destinationInput = new mapboxgl.Geocoder({
+      flyTo: false,
+      placeholder: 'Choose destination',
+      container: this.container.querySelector('#mapbox-directions-destination-input')
     });
 
-    this.destinationClear.addEventListener('click', () => {
-      this.destinationClear.classList.remove('active');
-      clearDestination();
+    this.map.addControl(this.destinationInput);
+
+    this.originInput.on('result', (e) => {
+      const coords = e.result.center;
+      createOrigin(coords);
+      this.animateToCoordinates('origin', coords);
     });
 
-    this.destinationInput.addEventListener('keydown', debounce((e) => {
-      if (this.nonKeyList(e.keyCode)) return;
-      const v = e.target.value.trim();
-      if (v.length > 2) {
-        queryDestination(e.target.value);
-      } else {
-        this.destinationClear.classList.remove('active');
-      }
-    }), 300);
+    this.originInput.on('clear', clearOrigin);
 
-    this.destinationInput.addEventListener('change', () => {
-      if (this.destinationTypeahead.selected) {
-        const coords = this.destinationTypeahead.selected.center;
-        destinationCoordinates(coords);
-        this.animateToCoordinates('destination', coords);
-      }
+    this.destinationInput.on('result', (e) => {
+      const coords = e.result.center;
+      createDestination(coords);
+      this.animateToCoordinates('destination', coords);
     });
 
-    this.originTypeahead = new typeahead(this.originInput, []);
-    this.destinationTypeahead = new typeahead(this.destinationInput, []);
-
-    // Filter results by place_name
-    this.originTypeahead.getItemValue = function(item) { return item.place_name; };
-    this.destinationTypeahead.getItemValue = function(item) { return item.place_name; };
+    this.destinationInput.on('clear', clearDestination);
 
     // Driving / Walking / Cycling profiles
     const profiles = this.container.querySelectorAll('input[type="radio"]');
@@ -144,72 +106,44 @@ export default class Inputs {
     // Reversing Origin / Destination
     this.container
       .querySelector('.js-reverse-inputs')
-      .addEventListener('click', reverse);
+      .addEventListener('click', () => {
+        const { origin, destination } = this.store.getState();
+        if (origin) this.actions.queryDestination(origin.geometry.coordinates);
+        if (destination) this.actions.queryOrigin(destination.geometry.coordinates);
+        reverse();
+      });
   }
 
   render() {
     this.store.subscribe(() => {
       const {
         originQuery,
-        originLoading,
-        originResults,
         destinationQuery,
-        destinationLoading,
-        destinationResults
+        originQueryCoordinates,
+        destinationQueryCoordinates
       } = this.store.getState();
 
-      if (!originResults.length || !originQuery.length) {
-        this.originTypeahead.clear();
-      }
-
-      if (!destinationResults.length || !destinationQuery.length) {
-        this.destinationTypeahead.clear();
-      }
-
-      this.originTypeahead.update(originResults);
-      this.destinationTypeahead.update(destinationResults);
-
       if (originQuery) {
-        this.originClear.classList.add('active');
-      } else {
-        this.originClear.classList.remove('active');
-      }
-
-      if (originLoading) {
-        this.originLoading.classList.add('active');
-      } else {
-        this.originLoading.classList.remove('active');
+        this.originInput.query(originQuery);
+        this.actions.queryOrigin(null);
       }
 
       if (destinationQuery) {
-        this.destinationClear.classList.add('active');
-      } else {
-        this.destinationClear.classList.remove('active');
+        this.destinationInput.query(destinationQuery);
+        this.actions.queryDestination(null);
       }
 
-      if (destinationLoading) {
-        this.destinationLoading.classList.add('active');
-      } else {
-        this.destinationLoading.classList.remove('active');
+      if (originQueryCoordinates) {
+        this.originInput.setInput(originQueryCoordinates);
+        this.animateToCoordinates('origin', originQueryCoordinates);
+        this.actions.queryOriginCoordinates(null);
       }
 
-      var onChange = document.createEvent('HTMLEvents');
-      onChange.initEvent('change', true, false);
-
-      // Adjust values if input is not :focus
-      // or query remains unchanged.
-      if (this.originInput !== document.activeElement &&
-          this.originInput.value !== originQuery) {
-        this.originInput.value = originQuery;
-        this.originInput.dispatchEvent(onChange);
+      if (destinationQueryCoordinates) {
+        this.destinationInput.setInput(destinationQueryCoordinates);
+        this.animateToCoordinates('destination', destinationQueryCoordinates);
+        this.actions.queryDestinationCoordinates(null);
       }
-
-      if (this.destinationInput !== document.activeElement &&
-          this.destinationInput.value !== destinationQuery) {
-        this.destinationInput.value = destinationQuery;
-        this.destinationInput.dispatchEvent(onChange);
-      }
-
     });
   }
 }